@mhvk 2018-07-01
Ideally, the whole initializer is rewritten, but pending that I think being "lazy" is a good idea!