@astrofrog 2018-07-01
@mhvk - yeah, this is one of these cases where I'd rather just make such a simple fix in case the refactoring never happens :)