@mhvk 2018-07-01
I'll wait a bit with merging just so @adrn and @eteq can have a look.