@eteq 2018-07-12
I'm loving the simplicity!  LGTM, so I'm merging.